Moved charger delete to edit panel
