+++ react-app/src/components/pages/chargers/page-components/ChargerEditPanel.tsx
@@ -1,9 +1,9 @@
 import { Charger } from '@/remote-access/types';
-import { AppBar, Box, Button, Divider, Grid, IconButton, LinearProgress, Paper, Theme, Toolbar, Typography } from '@material-ui/core';
-import React, { FC, useEffect, useState } from 'react';
+import { AppBar, Box, Button, Dialog, DialogActions, DialogTitle, Divider, IconButton, LinearProgress, Paper, Theme, Toolbar, Typography } from '@material-ui/core';
+import React, { FC, FormEvent, useEffect, useState } from 'react';
 import { chargerCollection } from '@/remote-access';
 import { useTheme } from '@material-ui/styles';
-import { Close } from '@material-ui/icons';
+import { Close, Delete } from '@material-ui/icons';
 import { Alert, ToggleButton, ToggleButtonGroup } from '@material-ui/lab';
 
 interface ChargerEditPanelProps {
@@ -37,6 +37,21 @@ const ChargerEditPanel: FC<ChargerEditPanelProps> = ({ chargerID }) => {
     setAvailability(newAvailability);
   };
 
+  const [openDelete, setDeleteOpen] = useState(false);
+  const handleDeleteClicked = () => {
+    setDeleteOpen(true);
+  }; 
+  
+  const handleCloseDelete = () => {
+    setDeleteOpen(false);
+  };
+
+  const onDeleteCharger = async (event: FormEvent<HTMLFormElement>, chargerId: number) => {
+    event.preventDefault();
+    const [res, error] = await chargerCollection.deleteChargerById(chargerId);
+    console.log(res, error);
+  };
+
   return (
     <>
       <Paper component="aside">
@@ -55,7 +70,7 @@ const ChargerEditPanel: FC<ChargerEditPanelProps> = ({ chargerID }) => {
             >
               <Toolbar variant="dense">
                 <Typography style={{ flexGrow: 1 }} variant="h5">
-                  Station Info
+                  Charger Info
                 </Typography>
                 <IconButton
                   aria-label="deselect charger"
@@ -85,22 +100,31 @@ const ChargerEditPanel: FC<ChargerEditPanelProps> = ({ chargerID }) => {
             <Box 
               sx={{ px: 4 }}
               borderTop={1}
+              padding={2}
               borderColor="error.main"
             >
-              <Grid container>
-                <Grid item lg={8}>
-                  <Typography variant="caption">
-                    Delete this Station
-                    <br />
-                    A deleted station is marked as Inactive
-                  </Typography>
-                </Grid>
-                <Grid item xs={12} lg={4}>
-                  <Button fullWidth variant="contained">
-                    Delete
-                  </Button>
-                </Grid>
-              </Grid>
+              <Button
+                startIcon={<Delete />}
+                style={{ color: theme.flexiCharge.primary.white }}
+                variant="contained"
+                color="secondary"
+                onClick={() => handleDeleteClicked()}
+              >
+                Delete
+              </Button>
+              <Dialog open={openDelete} onClose={handleCloseDelete}>
+                <DialogTitle>Are you sure you want to delete charger {charger.chargerID}?</DialogTitle>
+                <DialogActions>
+                  <form onSubmit={(e) => { onDeleteCharger(e, charger.chargerID); }}>
+                    <Button type="button" autoFocus onClick={handleCloseDelete}>
+                      Cancel
+                    </Button>
+                    <Button type="submit" variant="contained" color="secondary" startIcon={<Delete />} >
+                      Delete
+                    </Button>
+                  </form>
+                </DialogActions>
+              </Dialog>
             </Box>
           </>
         }
