+++ react-app/src/components/pages/chargers/page-components/ChargerRow.tsx
@@ -1,11 +1,10 @@
 import { Charger } from '@/remote-access/types';
 import {
   Theme, useTheme, TableRow, TableCell, Checkbox, Box, Typography,
-  Dialog, DialogTitle, Button, DialogActions
+  Button 
 } from '@material-ui/core';
-import { Delete, Error, Edit } from '@material-ui/icons';
-import { chargerCollection } from '@/remote-access';
-import React, { FC, FormEvent, useState } from 'react';
+import { Error, Edit } from '@material-ui/icons';
+import React, { FC } from 'react';
 
 interface ChargerRowProps {
   editClicked: (chargerID: number) => void
@@ -16,21 +15,6 @@ interface ChargerRowProps {
 
 const ChargerRow: FC<ChargerRowProps> = ({ charger, classes, editClicked, deleteClicked }) => {
   const theme: Theme = useTheme();
-
-  const [openDelete, setDeleteOpen] = useState(false);
-  const handleDeleteClicked = () => {
-    setDeleteOpen(true);
-  }; 
-  
-  const handleCloseDelete = () => {
-    setDeleteOpen(false);
-  };
-
-  const onDeleteCharger = async (event: FormEvent<HTMLFormElement>, chargerId: number) => {
-    event.preventDefault();
-    const [res, error] = await chargerCollection.deleteChargerById(chargerId);
-    console.log(res, error);
-  };
 
   return (
     <>
@@ -69,28 +53,6 @@ const ChargerRow: FC<ChargerRowProps> = ({ charger, classes, editClicked, delete
           >
             Edit
           </Button>
-          <Button
-            startIcon={<Delete />}
-            style={{ color: theme.flexiCharge.primary.white }}
-            variant="contained"
-            color="secondary"
-            onClick={() => handleDeleteClicked()}
-          >
-            Delete
-          </Button>
-          <Dialog open={openDelete} onClose={handleCloseDelete}>
-            <DialogTitle>Are you sure you want to delete charger {charger.chargerID}?</DialogTitle>
-            <DialogActions>
-              <form onSubmit={(e) => { onDeleteCharger(e, charger.chargerID); }}>
-                <Button type="button" autoFocus onClick={handleCloseDelete}>
-                  Cancel
-                </Button>
-                <Button type="submit" variant="contained" color="secondary" className={classes.dialogDelete} >
-                  Delete
-                </Button>
-              </form>
-            </DialogActions>
-          </Dialog>
         </TableCell>
       </TableRow>
     </>
